requirements.txt
streamlit 
pandas 
google-generativeai 
matplotlib 
plotly 
wordcloud 
openpyxl